requirements.txt
Pillow==9.2.0
base58==2.1.1
solana==0.25.1